fairymax: fix install location of shared files
As written, the formula for fairymax installs shared files to an
incorrect location. One other Homebrew formula uses fairymax as a
dependency: xboard. That formula is unable to locate ini files installed
by fairymax because they have been installed in a non-standard location.

To replicate this issue, install the xboard formula and run it using the
`xboard` command. You should notice an error message about being unable
to locate `fmax.ini`.

The standard location for the shared files installed by fairymax:

"$(brew --prefix)/share/games/fairymax/"

The current location where these shared files are installed:

"$(brew --prefix)/share/fairymax/"

The change in this commit simply passes the correct prefix to the
Makefile to ensure that all resources are placed in the correct
locations.

Additionally, as the copyright text conveys this is a public domain
work, the appropriate license type has been added to the formula.